He.RefGraph

Documentation

data RefGraph a

emptyRefGraph :: Ord a => RefGraph a

nodeLens :: Ord a => a -> Lens' (RefGraph a) (Maybe (Set a))

addDef :: Ord a => a -> RefGraph a -> RefGraph a

addRef :: Ord a => a -> a -> RefGraph a -> RefGraph a

type RefGraphT k m = StateT (RefGraph k) m

runRefGraphT :: (Ord k, Monad m) => RefGraphT k m a -> m (a, [SCC k])

addDefM :: (Ord k, Monad m) => k -> RefGraphT k m ()

addRefM :: (Ord k, Monad m) => k -> k -> RefGraphT k m ()

toGraph :: Ord a => RefGraph a -> Graph

toEdgeList :: Ord a => RefGraph a -> EdgeList a

filter :: (a -> Bool) -> RefGraph a -> RefGraph a

isCycle :: SCC a -> Bool

sccs :: Ord a => RefGraph a -> [SCC a]

data SCC vertex :: * -> *

Strongly connected component.

Constructors

AcyclicSCC vertex
A single vertex that is not in any cycle.

CyclicSCC [vertex]
A maximal set of mutually reachable vertices.

flattenSCC :: SCC vertex -> [vertex]

The vertices of a strongly connected component.

flattenSCCs :: [SCC a] -> [a]

The vertices of a list of strongly connected components.